fix SurValue save

+++ src/main/java/com/nn/dto/SurValueDto.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public record SurValueDto(
         String company,
         String chdrnum,
-        float surrenderValue,
+        Double surrenderValue,
         String jobUser,
         String jobName,
         LocalDate jobTimestamp) {
